Pre-check custom socket settings

+++ scripts/docker-entrypoint.sh
@@ -183,16 +183,23 @@ set_mysql_custom_settings "client" "socket" "MYSQL_SOCKET_DIR" "/mysqld.sock" "$
 set_mysql_custom_settings "mysql"  "socket" "MYSQL_SOCKET_DIR" "/mysqld.sock" "${DB_CUSTOM_CONF_DIR}/socket.cnf"
 set_mysql_custom_settings "mysqld" "socket" "MYSQL_SOCKET_DIR" "/mysqld.sock" "${DB_CUSTOM_CONF_DIR}/socket.cnf"
 
-# Create socket directory
-if [ ! -d "${MYSQL_SOCKET_DIR}" ]; then
-	run "mkdir -p ${MYSQL_SOCKET_DIR}"
-# Delete existing socket file
-elif [ -f "${MYSQL_SOCKET_DIR}/mysqld.sock" ]; then
-	run "rm -f ${MYSQL_SOCKET_DIR}/mysqld.sock"
+# Take care about custom socket directories
+if set | grep "^MYSQL_SOCKET_DIR=" >/dev/null 2>&1; then
+
+	# Create socket directory
+	if [ ! -d "${MYSQL_SOCKET_DIR}" ]; then
+		run "mkdir -p ${MYSQL_SOCKET_DIR}"
+
+	# Delete existing socket file
+	elif [ -f "${MYSQL_SOCKET_DIR}/mysqld.sock" ]; then
+		run "rm -f ${MYSQL_SOCKET_DIR}/mysqld.sock"
+	fi
+
+	# Set socket permission
+	run "chown ${DB_USER}:${DB_GROUP} ${MYSQL_SOCKET_DIR}"
+	run "chmod 777 ${MYSQL_SOCKET_DIR}"
 fi
-# Set socket permission
-run "chown ${DB_USER}:${DB_GROUP} ${MYSQL_SOCKET_DIR}"
-run "chmod 777 ${MYSQL_SOCKET_DIR}"
+
